+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * references in LaTeX IGT examples
+* proper handling of glossing abbreviations in LaTeX output
 
 ### Removed